model_espnet/requirements.txt
absl-py==0.12.0
aniso8601==8.1.0
appdirs==1.4.4
apturl==0.5.2
audioread==2.1.9
bcrypt==3.1.7
blinker==1.4
Bottleneck==1.3.2
Brlapi==0.7.0
cachetools==4.2.2
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
ci-sdr==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==1.4.1
configparser==5.0.2
cryptography==2.8
ctc-segmentation==1.6.6
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.23
dbus-python==1.2.16
decorator==5.0.9
defer==1.0.6
Distance==0.1.3
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.0.0
docker-pycreds==0.4.0
duplicity==0.8.12.0
editdistance==0.5.2
einops==0.3.0
entrypoints==0.3
espnet==0.9.10
espnet-tts-frontend==0.0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
future==0.18.2
g2p-en==2.1.0
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.30.2
google-auth-oauthlib==0.4.4
grpcio==1.38.0
gunicorn==20.0.4
h5py==3.2.1
httplib2==0.14.0
humanfriendly==9.1
idna==2.8
inflect==5.3.0
itsdangerous==1.1.0
jaconv==0.3
Jinja2==2.11.2
joblib==1.0.1
kaldiio==2.17.2
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
librosa==0.8.1
llvmlite==0.36.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.1.0
monotonic==1.5
nara-wpe==0.0.7
netifaces==0.10.4
nltk==3.6.2
numba==0.53.1
numpy==1.20.3
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.6.0
pathtools==0.1.2
pexpect==4.6.0
Pillow==7.0.0
pipenv==2020.11.15
pooch==1.4.0
promise==2.3
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.11.2
PyNaCl==1.3.0
pyparsing==2.4.7
pypinyin==0.41.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytorch-wpe==0.0.1
pytz==2019.3
pyworld==0.3.0
pyxdg==0.26
PyYAML==5.3.1
regex==2021.4.4
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
SecretStorage==2.3.1
sentencepiece==0.1.86
sentry-sdk==1.1.0
shortuuid==1.0.1
simplejson==3.16.0
six==1.14.0
smmap==4.0.0
SoundFile==0.10.3.post1
subprocess32==3.5.4
systemd-python==234
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
threadpoolctl==2.1.0
torch==1.8.1
torch-complex==0.2.1
tqdm==4.61.0
typeguard==2.12.1
typing-extensions==3.10.0.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.2.0
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.2.2
virtualenv-clone==0.5.4
wadllib==1.3.3
wandb==0.10.31
Werkzeug==1.0.1
xkit==0.0.0
